(BSR)[PRO] refactor: remove index files
