+++ pro/src/screens/Bookings/BookingsRecapTable/BookingsTable/CollectiveActionButtons.module.scss
