--- pro/src/screens/Bookings/BookingsRecapTable/BookingsTable/CollectiveActionButtons/index.ts
@@ -1,1 +1,0 @@
-export { default } from './CollectiveActionButtons'
