+++ pro/src/screens/Bookings/BookingsRecapTable/BookingsTable/CollectiveBookingDetails.module.scss
