--- pro/src/screens/Bookings/BookingsRecapTable/BookingsTable/CollectiveBookingDetails/index.ts
@@ -1,1 +1,0 @@
-export { default } from './CollectiveBookingDetails'
