+++ pro/src/screens/Bookings/BookingsRecapTable/BookingsTable/CollectiveBookingsTable.tsx
@@ -16,7 +16,7 @@ import {
 } from 'screens/Bookings/BookingsRecapTable/utils/sortingFunctions'
 import { Pagination } from 'ui-kit/Pagination'
 
-import { FilterByBookingStatus } from '../Filters'
+import { FilterByBookingStatus } from '../Filters/FilterByBookingStatus'
 import { NoFilteredBookings } from '../NoFilteredBookings/NoFilteredBookings'
 
 import styles from './BookingsTable.module.scss'
