--- pro/src/screens/Bookings/BookingsRecapTable/BookingsTable/CollectiveTimeLine/index.ts
@@ -1,1 +1,0 @@
-export { default } from './CollectiveTimeLine'
