+++ pro/src/screens/Bookings/BookingsRecapTable/BookingsTable/__specs__/CollectiveActionButtons.spec.tsx
@@ -9,8 +9,10 @@ import {
 } from 'utils/collectiveApiFactories'
 import { renderWithProviders } from 'utils/renderWithProviders'
 
-import CollectiveActionButtons from '..'
-import { CollectiveActionButtonsProps } from '../CollectiveActionButtons'
+import {
+  CollectiveActionButtons,
+  CollectiveActionButtonsProps,
+} from '../CollectiveActionButtons'
 
 const renderCollectiveActionButtons = (props: CollectiveActionButtonsProps) => {
   renderWithProviders(<CollectiveActionButtons {...props} />)
