--- pro/src/screens/Bookings/BookingsRecapTable/Filters/FilterByBookingStatus/index.ts
@@ -1,1 +1,0 @@
-export { default } from './FilterByBookingStatus'
