+++ pro/src/screens/Bookings/BookingsRecapTable/Filters/FilterByOmniSearch.module.scss
