--- pro/src/screens/Bookings/BookingsRecapTable/Filters/FilterByOmniSearch/index.ts
@@ -1,1 +1,0 @@
-export { default } from './FilterByOmniSearch'
