--- pro/src/screens/Bookings/BookingsRecapTable/Filters/_constants.ts
@@ -1,3 +1,0 @@
-export const ALL_BOOKING_STATUS = []
-export const DEFAULT_OMNISEARCH_CRITERIA = 'offre'
-export const EMPTY_FILTER_VALUE = ''
