+++ pro/src/screens/Bookings/BookingsRecapTable/Filters/constants.ts
@@ -1,5 +1,9 @@
 import { BookingOmniSearchFilters } from './types'
 
+export const ALL_BOOKING_STATUS = []
+export const DEFAULT_OMNISEARCH_CRITERIA = 'offre'
+export const EMPTY_FILTER_VALUE = ''
+
 type OmnisearchFilter = {
   id: string
   placeholderText: string
