--- pro/src/screens/Bookings/BookingsRecapTable/Filters/index.ts
@@ -1,3 +1,0 @@
-export { default as FilterByBookingStatus } from './FilterByBookingStatus'
-export { default as FilterByOmniSearch } from './FilterByOmniSearch'
-export * from './_constants'
