+++ pro/src/screens/Bookings/BookingsRecapTable/Header/__specs__/Header.spec.tsx
@@ -1,7 +1,7 @@
 import { render, screen } from '@testing-library/react'
 import React from 'react'
 
-import Header, { HeaderProps } from '../Header'
+import { Header, HeaderProps } from '../Header'
 
 const renderHeader = (props: HeaderProps) => render(<Header {...props} />)
