--- pro/src/screens/Bookings/BookingsRecapTable/index.ts
@@ -1,1 +1,0 @@
-export { default } from './BookingsRecapTable'
